Add team info test
