--- test/may_not_need_test.rb
@@ -1,34 +1,0 @@
-# may not need this anymore
-
-# require_relative 'test_helper'
-# require 'csv'
-# require 'minitest/autorun'
-# require 'minitest/pride'
-# require './lib/team_collection'
-# require './lib/team'
-#
-# class TeamCollectionTest < Minitest::Test
-#   def setup
-#     @team_collection = TeamCollection.new("./data/teams.csv")
-#     @team = @team_collection.teams.first
-#   end
-#
-#   def test_it_exists
-#     assert_instance_of TeamCollection, @team_collection
-#   end
-#
-#   def test_it_can_create_teams_from_csv
-#     assert_instance_of Team, @team
-#     assert_equal 1, @team.team_id
-#     assert_equal "Atlanta United", @team.team_name
-#   end
-#
-#   def test_it_has_team_info
-#     expected = {:team_id => 1,
-#                 :franchise_id => 23,
-#                 :team_name => "Atlanta United",
-#                 :abbreviation => "ATL",
-#                 :link => "/api/v1/teams/1"}
-#     assert_equal expected, @team_collection.team_info(1)
-#   end
-# end
